Refactoring the code in packege-fields
